+++ org.eclipse.contribution.weaving.jdt.tests/src/org/eclipse/contribution/weaving/jdt/tests/WeavingTestCase.java
@@ -374,7 +374,7 @@ public class WeavingTestCase extends TestCase {
        boolean errorFound = false;
        sb.append("Problems:\n");
      for (IMarker marker : markers) {
-       if (((Integer) marker.getAttribute(IMarker.SEVERITY)).intValue() == IMarker.SEVERITY_ERROR) {
+       if ((Integer) marker.getAttribute(IMarker.SEVERITY) == IMarker.SEVERITY_ERROR) {
          sb.append("  ");
          sb.append(marker.getResource().getName()).append(" : ");
          sb.append(marker.getAttribute(IMarker.LINE_NUMBER)).append(" : ");
